fix: Collect the correct PSNR key + refactor
The updated libvmaf now outputs multiple `psnr_` values, one for each channel component of the YCbCr colorspace. Originally this seems to only have been for `psnr_y`.

The current logic finds a metric that matches `psnr` in the key which ends up being `psnr_cr`, this is less useful and has been corrected to `psnr_y`.

The logic here was repeated for each metric type, I've DRY'd that up to leverage the new `get_metrics_list()` utility and a dictionary to lookup the corresponding JSON key for each metric type. Similar was done at the end for the `table_title`.